Microservices/face-service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
deepface==0.0.93
psycopg2-binary==2.9.9
psycopg-binary==3.2.9
sqlalchemy==2.0.35
nats-py==2.9.0
python-multipart==0.0.12
numpy>=1.23.5,<2.0.0
pgvector==0.3.0
python-dotenv==1.0.1
pyyaml==6.0.2
tensorflow==2.16.1
opencv-python==4.10.0.84
aiofiles==23.1.0
pydantic==2.7.3
ml-dtypes==0.3.1
tensorboard>=2.16,<2.17
Flask>=1.1.2
flask-cors>=4.0.1
gunicorn>=20.1.0
beautifulsoup4>=4.12.0
tf-keras>=2.16.0